+++ src/backend/mod.rs
@@ -298,10 +298,10 @@ pub trait Backend: Debug + Send + Sync {
             None => {
                 let installed_symlink = self.fa().installs_path.join("latest");
                 if installed_symlink.exists() {
-                    if !installed_symlink.is_symlink() {
+                    if installed_symlink.is_dir() && !installed_symlink.is_symlink() {
                         return Ok(Some("latest".to_string()));
                     }
-                    let target = installed_symlink.read_link()?;
+                    let target = file::resolve_symlink(&installed_symlink)?;
                     let version = target
                         .file_name()
                         .ok_or_else(|| eyre!("Invalid symlink target"))?