+++ src/backend/npm.rs
@@ -1,12 +1,11 @@
-use std::fmt::Debug;
-
 use serde_json::Value;
+use std::fmt::Debug;
 
 use crate::backend::{Backend, BackendType};
 use crate::cache::CacheManager;
 use crate::cli::args::BackendArg;
 use crate::cmd::CmdLineRunner;
-use crate::config::{Config, Settings};
+use crate::config::Config;
 use crate::env;
 use crate::install_context::InstallContext;
 use crate::toolset::ToolRequest;
@@ -59,8 +58,6 @@ impl Backend for NPMBackend {
 
     fn install_version_impl(&self, ctx: &InstallContext) -> eyre::Result<()> {
         let config = Config::try_get()?;
-        let settings = Settings::get();
-        settings.ensure_experimental("npm backend")?;
 
         CmdLineRunner::new("npm")
             .arg("install")
@@ -75,6 +72,14 @@ impl Backend for NPMBackend {
             .execute()?;
 
         Ok(())
+    }
+
+    #[cfg(windows)]
+    fn list_bin_paths(
+        &self,
+        tv: &crate::toolset::ToolVersion,
+    ) -> eyre::Result<Vec<std::path::PathBuf>> {
+        Ok(vec![tv.install_path()])
     }
 }