+++ src/cli/settings/ls.rs
@@ -58,13 +58,14 @@ mod tests {
     #[test]
     fn test_settings_ls() {
         reset();
-        assert_cli_snapshot!("settings", @r###"
+        assert_cli_snapshot!("settings", @r#"
         activate_aggressive = false
         all_compile = false
         always_keep_download = true
         always_keep_install = true
         asdf = true
         asdf_compat = false
+        cache_prune_age = "0"
         cargo_binstall = true
         color = true
         disable_default_shorthands = false
@@ -97,23 +98,30 @@ mod tests {
         vfox = false
         yes = true
 
+        [ruby]
+        default_packages_file = "~/.default-gems"
+        ruby_build_repo = "https://github.com/rbenv/ruby-build.git"
+        ruby_install = false
+        ruby_install_repo = "https://github.com/postmodern/ruby-install.git"
+
         [status]
         missing_tools = "if_other_versions_installed"
         show_env = false
         show_tools = false
-        "###);
+        "#);
     }
 
     #[test]
     fn test_settings_ls_keys() {
         reset();
-        assert_cli_snapshot!("settings", "--keys", @r###"
+        assert_cli_snapshot!("settings", "--keys", @r#"
         activate_aggressive
         all_compile
         always_keep_download
         always_keep_install
         asdf
         asdf_compat
+        cache_prune_age
         cargo_binstall
         color
         disable_default_shorthands
@@ -140,6 +148,11 @@ mod tests {
         python_pyenv_repo
         quiet
         raw
+        ruby
+        ruby.default_packages_file
+        ruby.ruby_build_repo
+        ruby.ruby_install
+        ruby.ruby_install_repo
         status
         status.missing_tools
         status.show_env
@@ -149,6 +162,6 @@ mod tests {
         verbose
         vfox
         yes
-        "###);
+        "#);
     }
 }